+++ rbx.el
@@ -5,7 +5,7 @@
 ;; Author: rbx-for-emacs contributors
 ;; Maintainer: rbx-for-emacs contributors
 ;; Version: 0.1.0
-;; Package-Requires: ((emacs "30.1") (magit-section "4.1.0") (transient "0.7.5") (yaml "1.2.0"))
+;; Package-Requires: ((emacs "30.1") (magit-section "4.1.0") (transient "0.7.5"))
 ;; Keywords: tools, languages
 ;; URL: https://github.com/luishgh/rbx-for-emacs